refactor(storage): extract the shared WAL sidecar freeze helper
Five crates each carried a byte-identical copy of the fixture that makes a
SQLite database's -wal/-shm sidecars read-only, and thirteen further call sites
wrote the same sequence inline. Move it into a single `test_support` module in
khive-storage behind an off-by-default `test-support` feature, and have every
caller use it.

khive-storage is the home because it is the only crate every consumer depends
on. khive-db reads as the more natural place given the helper's subject, but
khive-pack-knowledge does not depend on khive-db, so a helper there could not
reach two of the inline sites. A `#[cfg(test)]` module cannot cross crate
boundaries at all, which is why this is a feature rather than a test-cfg.

Behaviour is unchanged; the helper body is the existing one.
