+++ crates/khive-mcp/Cargo.toml
@@ -82,6 +82,7 @@ test-forward-seam = []
 [dev-dependencies]
 tokio = { workspace = true, features = ["test-util"] }
 khive-runtime = { version = "0.8.0", path = "../khive-runtime", features = ["fault-injection"] }
+khive-storage = { version = "0.8.0", path = "../khive-storage", features = ["test-support"] }
 rmcp = { workspace = true, features = ["server", "transport-io", "client"] }
 async-trait = { workspace = true }
 serial_test = { workspace = true }
