+++ crates/khive-runtime/Cargo.toml
@@ -47,6 +47,7 @@ sha2 = { workspace = true }
 
 [dev-dependencies]
 khive-gate-rego = { version = "0.8.0", path = "../khive-gate-rego" }
+khive-storage = { version = "0.8.0", path = "../khive-storage", features = ["test-support"] }
 tokio = { workspace = true, features = ["test-util"] }
 tempfile = { workspace = true }
 tracing-subscriber = { workspace = true }
