+++ crates/khive-storage/Cargo.toml
@@ -23,3 +23,6 @@ uuid = { workspace = true }
 
 [dev-dependencies]
 tokio = { workspace = true }
+
+[features]
+test-support = []
