+++ crates/kkernel/src/code_audit.rs
@@ -1696,21 +1696,7 @@ crate-b = 1
     }
 
     #[cfg(unix)]
-    fn freeze_snapshot_sidecars(path: &Path) {
-        use std::os::unix::fs::PermissionsExt;
-        for suffix in ["-wal", "-shm"] {
-            let mut name = path.file_name().expect("db file name").to_os_string();
-            name.push(suffix);
-            let sidecar = path.parent().expect("db parent dir").join(name);
-            if sidecar.exists() {
-                let mut permissions = std::fs::metadata(&sidecar)
-                    .expect("sidecar metadata")
-                    .permissions();
-                permissions.set_mode(0o444);
-                std::fs::set_permissions(&sidecar, permissions).expect("freeze sidecar");
-            }
-        }
-    }
+    use khive_storage::test_support::freeze_snapshot_sidecars;
 
     /// `generate_report` behind the frozen-snapshot form its read-only open
     /// accepts. SQLite connection close is deferred, so a fixture seeded and
